# docs/openapi.json
{
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "examples": [
          {
            "error": "Bad Request",
            "message": "Invalid input data.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message",
          "error"
        ],
        "type": "object"
      },
...
      "InternalServerError": {
        "description": "Internal Server Error",
        "examples": [
          {
            "error": "Internal Server Error.",
            "message": "An unexpected error occurred.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Internal Server Error.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "An unexpected error occurred.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to a server error",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "NotFound": {
        "description": "Not Found",
        "examples": [
          {
            "error": "Not Found.",
            "message": "Resource not found.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Not Found.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "Resource not found.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the requested resource could not be found",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "examples": [
          {
            "error": "Unauthorized.",
            "message": "Unauthorized access.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Unauthorized.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "Unauthorized access.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to authentication failure",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "schemas": {
...
      "ListTimeEntriesResponse": {
    "type": "object",
    "description": "Paginated response for listing time entries.",
    "additionalProperties": false,
    "properties": {
      "data": {
        "type": "array",
        "description": "List of time entries.",
        "items": {
          "$ref": "#/components/schemas/TimeEntryResponse"
        }
      },
      "meta": {
        "$ref": "#/components/schemas/Meta"
      }
    },
    "required": [
      "data",
      "meta"
    ]
  },
      "Meta": {
        "description": "Metadata information for paginated responses.",
        "example": {
          "limit": 10,
          "page": 1,
          "total": 9
        },
        "properties": {
          "limit": {
            "example": 10,
            "type": "integer"
          },
          "page": {
            "example": 1,
            "type": "integer"
          },
          "total": {
            "example": 9,
            "type": "integer"
          }
        },
        "title": "Meta",
        "type": "object"
      },
...
      "TimeEntryRequest": {
        "additionalProperties": false,
        "description": "Schema for creating or updating a time entry.",
        "examples": [
          {
            "date": "2025-02-13",
            "end_time": "17:00:00",
            "notes": "Worked on feature X",
            "start_time": "09:00:00"
          }
        ],
        "properties": {
          "date": {
            "description": "The date for the time entry.",
            "format": "date",
            "type": "string"
          },
          "end_time": {
            "description": "The end time for the work done.",
            "format": "time",
            "type": "string"
          },
          "notes": {
            "description": "Additional notes for the time entry.",
            "type": "string"
          },
          "start_time": {
            "description": "The start time for the work done.",
            "format": "time",
            "type": "string"
          }
        },
        "required": [
          "date",
          "start_time",
          "end_time"
        ],
        "title": "TimeEntryRequest",
        "type": "object"
      },
      "TimeEntryResponse": {
        "additionalProperties": false,
        "description": "Response schema for a time entry.",
        "examples": [
          {
            "created_at": "2025-02-13T10:00:00Z",
            "created_by": "user@example.com",
            "date": "2025-02-13",
            "end_time": "2025-02-13T17:00:00Z",
            "hours": 8,
            "id": "123e4567-e89b-12d3-a456-426614174000",
            "is_timecard_generated": true,
            "issue_id": "issue789",
            "notes": "Daily work log",
            "project_id": "abc123",
            "start_time": "2025-02-13T09:00:00Z",
            "updated_at": "2025-02-13T18:00:00Z"
          }
        ],
        "properties": {
          "created_at": {
            "description": "When the time entry was created.",
            "format": "date-time",
            "type": "string"
          },
          "created_by": {
            "description": "The user who created the time entry.",
            "format": "email",
            "type": "string"
          },
          "date": {
            "description": "The date of the time entry.",
            "format": "date",
            "type": "string"
          },
          "end_time": {
            "description": "The end time of the work done.",
            "format": "date-time",
            "type": "string"
          },
          "hours": {
            "description": "The number of hours worked.",
            "type": "number"
          },
          "id": {
            "description": "Unique identifier for the time entry.",
            "format": "uuid",
            "type": "string"
          },
          "is_timecard_generated": {
            "description": "Indicates if a time card has been generated.",
            "type": "boolean"
          },
          "issue_id": {
            "description": "The issue ID the entry belongs to.",
            "type": "string"
          },
          "notes": {
            "description": "Notes about the time entry.",
            "type": "string"
          },
          "project_id": {
            "description": "The project ID for the time entry.",
            "type": "string"
          },
          "start_time": {
            "description": "The start time of the work done.",
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "description": "When the time entry was last updated.",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "issue_id",
          "created_by",
          "date",
          "start_time",
          "end_time",
          "hours",
          "created_at",
          "updated_at"
        ],
        "title": "TimeEntryResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT Token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@projectmanagement.com",
      "name": "Project Management API Team",
      "url": "http://projectmanagement.com"
    },
    "description": "API for managing clients, projects, and issues to demonstrate the features of OpenAPI 3.1 specification.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://example.com/terms/",
    "title": "Project Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/project/{project_id}/issue/{issue_id}/time-entry": {
    "post": {
      "description": "Create a time entry for a specific issue within a project.",
      "operationId": "CreateIssueTimeEntry",
      "parameters": [
        {
          "description": "ID of the project",
          "in": "path",
          "name": "project_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "ID of the issue",
          "in": "path",
          "name": "issue_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TimeEntryRequest"
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "Time entry created successfully.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeEntryResponse"
              }
            }
          }
        },
        "400": {
          "description": "Invalid request parameters.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/responses/BadRequest"
              }
            }
          }
        },
        "401": {
          "description": "Unauthorized request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/responses/Unauthorized"
              }
            }
          }
        },
        "404": {
          "description": "Project or issue not found.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/responses/NotFound"
              }
            }
          }
        },
        "500": {
          "description": "Internal server error.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/responses/InternalServerError"
              }
            }
          }
        }
      },
      "summary": "Create a time entry for an issue",
      "tags": ["Issue Time Entry"]
    }
  },
  "/project/{project_id}/issue/{issue_id}/time-entries": {
    "get": {
      "description": "Retrieve a list of time entries for a specific issue.",
      "operationId": "ListIssueTimeEntries",
      "parameters": [
        {
          "description": "ID of the project",
          "in": "path",
          "name": "project_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "ID of the issue",
          "in": "path",
          "name": "issue_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "Page number for pagination",
          "in": "query",
          "name": "page",
          "schema": {
            "minimum": 1,
            "type": "integer"
          }
        },
        {
          "description": "Number of time entries per page",
          "in": "query",
          "name": "page_size",
          "schema": {
            "minimum": 1,
            "type": "integer"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "List of time entries retrieved successfully.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListTimeEntriesResponse"
              }
            }
          }
        },
        "400": {
          "description": "Invalid request parameters."
        },
        "401": {
          "description": "Unauthorized request."
        },
        "404": {
          "description": "Project or issue not found."
        },
        "500": {
          "description": "Internal server error."
        }
      },
      "summary": "List time entries for an issue",
      "tags": ["Issue Time Entry"]
    }
  },
  "/project/{project_id}/issue/{issue_id}/time-entry/{te_id}": {
    "get": {
      "description": "Retrieve a specific time entry for an issue.",
      "operationId": "GetIssueTimeEntryByID",
      "parameters": [
        {
          "description": "ID of the project",
          "in": "path",
          "name": "project_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "ID of the issue",
          "in": "path",
          "name": "issue_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "ID of the time entry",
          "in": "path",
          "name": "te_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "Time entry retrieved successfully."
        },
        "404": {
          "description": "Time entry not found."
        }
      },
      "summary": "Get a specific time entry for an issue",
      "tags": ["Issue Time Entry"]
    },
    "put": {
      "description": "Update a specific time entry for an issue.",
      "operationId": "UpdateIssueTimeEntryByID",
      "parameters": [
        {
          "description": "ID of the project",
          "in": "path",
          "name": "project_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "ID of the issue",
          "in": "path",
          "name": "issue_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "ID of the time entry",
          "in": "path",
          "name": "te_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TimeEntryRequest"
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Time entry updated successfully."
        },
        "404": {
          "description": "Time entry not found."
        }
      },
      "summary": "Update a specific time entry for an issue",
      "tags": ["Issue Time Entry"]
    },
    "delete": {
      "description": "Delete a specific time entry for an issue.",
      "operationId": "DeleteIssueTimeEntry",
      "parameters": [
        {
          "description": "ID of the project",
          "in": "path",
          "name": "project_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "ID of the issue",
          "in": "path",
          "name": "issue_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "ID of the time entry",
          "in": "path",
          "name": "te_id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "Time entry deleted successfully."
        },
        "404": {
          "description": "Time entry not found."
        }
      },
      "summary": "Delete a specific time entry for an issue",
      "tags": ["Issue Time Entry"]
    }
  },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Test server",
      "url": "https://test.projectmanagement.com/api/v1"
    },
    {
      "description": "Development server",
      "url": "https://dev.projectmanagement.com/api/v1"
    },
    {
      "description": "Production server",
      "url": "https://api.projectmanagement.com/api/v1"
    }
  ],
...
}